fix: KeyAttestation#key method blowing up with OpenSSL 1.0.2
While working with OpenSSL 1.0.2, some test were failing with the error:
'OpenSSL::PKey::RSA missing key'.
This error is caused because no exponent is passed to the method
OpenSSL::PKey::RSA#set_key in 'lib/tpm/t_public.rb'.
Given that we only support the default exponent as for now, we can just
pass it to the method and it will work.