refactor: remove dupe certificates before adding them to trust store
The method add_cert from OpenSSL:X509:Store raises an exception if you
try to add a cert that it's already present.
The problem is that there are duplicate certificates in
KeyAttestation::ROOT_CERTIFICATES, particularly these two:
Microsoft RootCA and QC RootCA.

To fix this, we remove the duplicate certificates (certificates with
same serial number) before adding them.