fix: race condition caused by `not_after` being too short
There seems to be a race condition between the not_after attribute in
the certificate and the execution.
The problem is that when we create a certificate, we make it that is
only valid for just 60 seconds. Therefore, if the execution takes
longer than 60 seconds for checking if the certificate is valid,
eventually the certificate will be considered invalid.

To fix this, I decided to double the time in which a certificate is
considered valid.