test: fix test cases randomly failing due to certificates being 'not yet valid'
Test are being flaky in Travis, randomly failing given that
the certificate is not yet valid.
It seems that when checking if the certificate is valid, comparing
Time.now with the attribute `not_before` of a certificate (that is also
initiated with Time.now) is sometimes returning an error due to the
difference being too short – it could be related with the comparison of
floats that are close to each other.

To fix this, we just moved the `not_before` a little back in time, in
order to make the difference larger when checking if it's valid.