tests: fix tests failing with OpenSSL v3
The issue is that we set up the certificate to have an invalid version
(version is `3` initially) but the certificate is later signed which, in
newer versions of OpenSSL, updates the version to a valid one (version
changes to be `2`).

This PR fixes the issue by changing the `version` _after_ the
certificate has been signed.